Fix Debian package.
dh_shlibdeps was failing due to the use of shared library
in a non standard-directory. Teach it where the plug-in are located.
It will only be required as long as the coshell is directly linked
against the dynamic graph plug-in.

Remove postinst/postrm scripts as debhelper puts a ldconfig call inside.
ldconfig is used to update the linker database of available shared libraries
in the standard directories (see man ld.so). It does not track plug-ins so
this is not needed and trigger a lintian warning.